[kernel] Implemented 64-bit offset support.
ISOShift now contains a 34-bit rshifted offset, which is converted to
ISOShift64 before use. (TODO: Store a 64-bit offset, though that may
require rearranging the memory map.)

DVD-9 multi-game disc images on exFAT partitions now work properly.

DVD-R5 and DVD-R9 discs still have some issues:
- My DVD-R5 with Super Mario Sunshine either crashes immediately after
  the "Progressive Scan" screen if I start it with SD saves, or on the
  title screen if I start it with USB saves.
- All games on my DVD-R9 complain that the disc cover is open.

+++ kernel/ISO.h
@@ -24,7 +24,7 @@ extern bool Datel;
 bool ISOInit();
 void ISOClose();
 void ISOSetupCache();
-const u8 *ISORead(u32* Length, u32 Offset);
-void ISOSeek(u32 Offset);
+const u8 *ISORead(u32* Length, u64 Offset);
+void ISOSeek(u64 Offset);
 
 #endif
